moved ChargerStationsTable to its own file
